collision detection

+++ src/utilz/HelpMethods.java
@@ -6,7 +6,7 @@ import main.Game;
  * HelpMethods
  */
 public class HelpMethods {
-  public static boolean CanMoveHere(float x, float y, int width, int height, int[][] lvlData) {
+  public static boolean CanMoveHere(float x, float y, float width, float height, int[][] lvlData) {
     if (!IsSolid(x, y, lvlData))
       if (!IsSolid(x + width, y + height, lvlData))
         if (!IsSolid(x + width, y, lvlData))
@@ -15,6 +15,8 @@ public class HelpMethods {
     return false;
   }
 
+  // IsSolid return true when player moves to solid area
+  //
   private static boolean IsSolid(float x, float y, int[][] lvlData) {
     if (x < 0 || x >= Game.GAME_WIDTH)
       return true;
@@ -24,6 +26,7 @@ public class HelpMethods {
     float yIndex = y / Game.TILES_SIZE;
     int value = lvlData[(int) yIndex][(int) xIndex];
 
+    // 11 is transparent image
     if (value >= 48 || value < 0 || value != 11)
       return true;
     return false;
